Expose the HTTPClient object.
Main use case for this is to turn on debugging so we can see what's
going between us and NationBuilder. However there are several options
that can be set on the object to tweak proxy and connection settings.

Also made some small changes to the test in order to make it easier to
grok. I moved the `NationBuilder::Client` object to use `subject` so it was
more clear when we're talking about the subject of the test (which is a
client) rather than the mocked `HTTPClient` client.