lib/nationbuilder: include response object in errors
This commit provides users with access to HTTP response objects
when the NationBuilder client encounters an error. Among other
things, this allows applications to respond intelligently to
different status codes. For example, checking for a 401 status
code can indicate to the host application that an access token
has expired or is otherwise invalid.

This commit is purely additive. Other than extending the error
objects with a `response` accessor, there should be no other
side effects.

Affected classes:

- ServerError
- ClientError
- RateLimitedError