Format query params to match expectations from the server
Per https://nationbuilder.com/people_api#search, the `people#search` 
endpoint supports searching by custom fields. However, using the client 
returns a 500 as the params are not being encoded as the server expects.

This adds a `Utils` module to address this to avoid introducing 
additional dependencies (this is included in Rack in 
`build_nested_query`, and ActiveSupport in Rails in `Hash.to_query`)

For context, several approaches are discussed in 
[StackOverflow](https://stackoverflow.com/questions/798710/ruby-how-to-turn-a-hash-into-http-parameters) 
- this follows what was proposed in [this blog 
post](https://mensfeld.pl/2012/01/converting-nested-hash-into-http-url-params-hash-version-in-ruby/), 
similar to [the QueryParam gem](https://github.com/simen/queryparams).